+++ Major/ViewController.swift
@@ -15,16 +15,16 @@ class ViewController: UIViewController {
     let sleepTime: UInt32 = 2
 
     @IBAction func singleAction(_ sender: UIButton) {
-        kommander.make { () -> TimeInterval in
+        kommander.do { () -> TimeInterval in
             sleep(self.sleepTime)
             return Date().timeIntervalSince1970
         }.success { result in
             print("Single: " + String(describing: result))
-        }.execute()
+        }.run()
     }
 
     @IBAction func concurrentAction(_ sender: UIButton) {
-        kommander.execute(kommander.make([ { () -> Any? in
+        kommander.run(kommander.do([ { () -> Any? in
             sleep(self.sleepTime)
             print("Concurrent first: " + String(describing: Date().timeIntervalSince1970))
             return nil
@@ -40,7 +40,7 @@ class ViewController: UIViewController {
     }
 
     @IBAction func sequentialAction(_ sender: UIButton) {
-        kommander.execute(kommander.make([ { () -> Any? in
+        kommander.run(kommander.do([ { () -> Any? in
             sleep(self.sleepTime)
             print("Sequential first: " + String(describing: Date().timeIntervalSince1970))
             return nil
@@ -56,18 +56,18 @@ class ViewController: UIViewController {
     }
 
     @IBAction func errorAction(_ sender: UIButton) {
-        kommander.make {
+        kommander.do {
             sleep(self.sleepTime)
             throw CocoaError(.featureUnsupported)
-        }.error { error in
+        }.error(CocoaError.self) { error in
             print("Error: " + String(describing: error!))
-        }.execute()
+        }.run()
     }
 
     @IBAction func crashAction(_ sender: UIButton) {
-        kommander.make {
+        kommander.do {
             sleep(self.sleepTime)
             fatalError()
-        }.execute()
+        }.run()
     }
 }