+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Release 1.1
+
+- [x] Set kommand error closure specifying Error type.
+- [x] Modern sugar syntax.
+
 # Release 1.0.2
 
 - [x] Ready for Swift 4.2.